simplifying callbacks, fixing poor job copying zvals
I've switched to using call_user_function_ex instead of zend_call_function, as it's a wrapper for it anyway.

This reduces a lof of the cognitive overhead in trying to set up the callbacks.

I also had copying over zvals incorrectly, so said anon functions not in a var were being destroyed, so they couldn't
be called in the actual callback. This should now be fixed