Reworking the callback for GearmanWorker
Note: There is still a memory leak. If you pass in an inline function from within
a class, there is a double refcount in the gearman callback obj free. But we need
the references when adding the function as well.

Whether or not you call work on the GearmanWorker object (as in running
gearman_worker_work) is agnostic to the memleak, which simplifies things.

Doing an inline function outside the context of a class does not reproduce the
memleak, which is something!