fixing add function for same job
I was treating the cb_list as an indexed hash and appending functions to it. Of course,
if you tried adding a callback to it twice that uses the same job name, it would get a
little confused. This should just update the currently added job with a new function