cleaning up (some) of test_client.php code
I'm actually deprecating this in favor of more paradigmatic tests in PHP. Right now, you have to run the entire suite here to get to any one test.

It makes it more difficult and allows for possible contamination in other tests. Boo.

That said, I might pull some of the code out of here for said proper tests. So it's good to have something working