Param to skip exception handling setup in addServer/addServers in GearmanClient class
Addressing issue #59 in wcgallego/pecl-gearman
(https://github.com/wcgallego/pecl-gearman/issues/59)

In 2013, a patch was introduced in the core libgearman code to do checks for
connections to the GearmanD daemon. In doing so, the libgearman function
gearman_client_set_server_option now does a connection to the GearmanD daemon,
breaking the contract that GearmanClient::addServer and addServers does not do
any connection work, but simply adds the host names to an array.

If the host cannot be reached, this will throw an exception.

This patch makes it optional to skip the exception handling code here and to
attempt making the connections as needed when actual work needs to be done.

Bumping to version 2.0.4